# packages/core/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ScoutGlobe API",
    "description": "Futbolcu verisi, transfer onerisi ve future-star kesfi icin API.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/players/search": {
      "get": {
        "tags": [
          "players"
        ],
        "summary": "Oyuncu ara ve filtrele",
        "operationId": "search_players_players_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 2
                },
                {
                  "type": "null"
                }
              ],
              "description": "Isim parcasi",
              "title": "Q"
            },
            "description": "Isim parcasi"
          },
          {
            "name": "position",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Pozisyon veya alt pozisyon",
              "title": "Position"
            },
            "description": "Pozisyon veya alt pozisyon"
          },
          {
            "name": "league_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "League Id"
            }
          },
          {
            "name": "country",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 2,
                  "maxLength": 2
                },
                {
                  "type": "null"
                }
              ],
              "title": "Country"
            }
          },
          {
            "name": "age_min",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 45,
                  "minimum": 14
                },
                {
                  "type": "null"
                }
              ],
              "title": "Age Min"
            }
          },
          {
            "name": "age_max",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 45,
                  "minimum": 14
                },
                {
                  "type": "null"
                }
              ],
              "title": "Age Max"
            }
          },
          {
            "name": "value_max",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Azami piyasa degeri (EUR)",
              "title": "Value Max"
            },
            "description": "Azami piyasa degeri (EUR)"
          },
          {
            "name": "minutes_min",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Son sezon asgari dakika",
              "title": "Minutes Min"
            },
            "description": "Son sezon asgari dakika"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerSearchResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/players/{player_id}/form": {
      "get": {
        "tags": [
          "players"
        ],
        "summary": "Mac bazli form egrisi ve sezon trendi",
        "operationId": "get_player_form_players__player_id__form_get",
        "parameters": [
          {
            "name": "player_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Player Id"
            }
          },
          {
            "name": "metric",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Metrik: minutes, goals, assists, goal_contributions",
              "default": "minutes",
              "title": "Metric"
            },
            "description": "Metrik: minutes, goals, assists, goal_contributions"
          },
          {
            "name": "window",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 20,
              "minimum": 1,
              "description": "Kayan ortalama penceresi (mac)",
              "default": 5,
              "title": "Window"
            },
            "description": "Kayan ortalama penceresi (mac)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 5,
              "description": "Kac maca bakilacak (en yeniden geriye)",
              "default": 60,
              "title": "Limit"
            },
            "description": "Kac maca bakilacak (en yeniden geriye)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerForm"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/players/{player_id}/shots": {
      "get": {
        "tags": [
          "players"
        ],
        "summary": "Sut haritasi ve bolge dagilimi",
        "operationId": "get_player_shots_players__player_id__shots_get",
        "parameters": [
          {
            "name": "player_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Player Id"
            }
          },
          {
            "name": "season",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Sezon, orn. 2025-26. Bos = tum sezonlar",
              "title": "Season"
            },
            "description": "Sezon, orn. 2025-26. Bos = tum sezonlar"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 10,
              "description": "Cizilecek azami sut sayisi",
              "default": 300,
              "title": "Limit"
            },
            "description": "Cizilecek azami sut sayisi"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerShots"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/players/{player_id}": {
      "get": {
        "tags": [
          "players"
        ],
        "summary": "Oyuncu profili",
        "operationId": "get_player_players__player_id__get",
        "parameters": [
          {
            "name": "player_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Player Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "FormSeries": {
        "properties": {
          "metric": {
            "type": "string",
            "title": "Metric"
          },
          "metricLabel": {
            "type": "string",
            "title": "Metriclabel"
          },
          "window": {
            "type": "integer",
            "title": "Window"
          },
          "totalMatches": {
            "type": "integer",
            "title": "Totalmatches"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/FormPoint"
            },
            "type": "array",
            "title": "Points"
          }
        },
        "type": "object",
        "required": [
          "metric",
          "metricLabel",
          "window",
          "totalMatches",
          "points"
        ],
        "title": "FormSeries"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlayerDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "fullName": {
            "type": "string",
            "title": "Fullname"
          },
          "imageUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Imageurl"
          },
          "position": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          },
          "subPosition": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subposition"
          },
          "birthDate": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birthdate"
          },
          "age": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age"
          },
          "nationalityCode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nationalitycode"
          },
          "clubId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clubid"
          },
          "clubName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clubname"
          },
          "leagueId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Leagueid"
          },
          "marketValueEur": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Marketvalueeur"
          },
          "clubLogoUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clublogourl"
          },
          "leagueName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Leaguename"
          },
          "leagueLogoUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Leaguelogourl"
          },
          "foot": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Foot"
          },
          "heightCm": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Heightcm"
          },
          "contractUntil": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contractuntil"
          },
          "seasonStats": {
            "items": {
              "$ref": "#/components/schemas/SeasonStatsOut"
            },
            "type": "array",
            "title": "Seasonstats",
            "default": []
          },
          "marketValueHistory": {
            "items": {
              "$ref": "#/components/schemas/MarketValuePoint"
            },
            "type": "array",
            "title": "Marketvaluehistory",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "fullName"
        ],
        "title": "PlayerDetail"
      },
      "PlayerForm": {
        "properties": {
          "playerId": {
            "type": "integer",
            "title": "Playerid"
          },
          "series": {
            "$ref": "#/components/schemas/FormSeries"
          },
          "seasons": {
            "items": {
              "$ref": "#/components/schemas/SeasonTrendPoint"
            },
            "type": "array",
            "title": "Seasons"
          }
        },
        "type": "object",
        "required": [
          "playerId",
          "series",
          "seasons"
        ],
        "title": "PlayerForm"
      },
      "PlayerSearchResult": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PlayerSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "PlayerSearchResult"
      },
      "PlayerShots": {
        "properties": {
          "playerId": {
            "type": "integer",
            "title": "Playerid"
          },
          "season": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Season"
          },
          "totalShots": {
            "type": "integer",
            "title": "Totalshots"
          },
          "totalGoals": {
            "type": "integer",
            "title": "Totalgoals"
          },
          "totalXg": {
            "type": "number",
            "title": "Totalxg"
          },
          "xgDifference": {
            "type": "number",
            "title": "Xgdifference"
          },
          "zones": {
            "items": {
              "$ref": "#/components/schemas/ShotZone"
            },
            "type": "array",
            "title": "Zones"
          },
          "shots": {
            "items": {
              "$ref": "#/components/schemas/ShotOut"
            },
            "type": "array",
            "title": "Shots"
          }
        },
        "type": "object",
        "required": [
          "playerId",
          "totalShots",
          "totalGoals",
          "totalXg",
          "xgDifference",
          "zones",
          "shots"
        ],
        "title": "PlayerShots"
      },
      "PlayerSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "fullName": {
            "type": "string",
            "title": "Fullname"
          },
          "imageUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Imageurl"
          },
          "position": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          },
          "subPosition": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subposition"
          },
          "birthDate": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birthdate"
          },
          "age": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age"
          },
          "nationalityCode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nationalitycode"
          },
          "clubId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clubid"
          },
          "clubName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clubname"
          },
          "leagueId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Leagueid"
          },
          "marketValueEur": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Marketvalueeur"
          }
        },
        "type": "object",
        "required": [
          "id",
          "fullName"
        ],
        "title": "PlayerSummary",
        "description": "Compact shape used in lists, squads and search results."
      },
...
      "SeasonTrendPoint": {
        "properties": {
          "season": {
            "type": "string",
            "title": "Season"
          },
          "matches": {
            "type": "integer",
            "title": "Matches"
          },
          "minutes": {
            "type": "integer",
            "title": "Minutes"
          },
          "minutesPerMatch": {
            "type": "number",
            "title": "Minutespermatch"
          },
          "goals": {
            "type": "integer",
            "title": "Goals"
          },
          "assists": {
            "type": "integer",
            "title": "Assists"
          },
          "goalsPer90": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Goalsper90"
          },
          "assistsPer90": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assistsper90"
          }
        },
        "type": "object",
        "required": [
          "season",
          "matches",
          "minutes",
          "minutesPerMatch",
          "goals",
          "assists"
        ],
        "title": "SeasonTrendPoint",
        "description": "Per-season aggregate — the \"is he growing\" view above the match noise."
      },
      "ShotOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "playedOn": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Playedon"
          },
          "minute": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Minute"
          },
          "xg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Xg"
          },
          "locationX": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locationx"
          },
          "locationY": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locationy"
          },
          "bodyPart": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bodypart"
          },
          "situation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Situation"
          },
          "result": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "isGoal": {
            "type": "boolean",
            "title": "Isgoal"
          }
        },
        "type": "object",
        "required": [
          "id",
          "isGoal"
        ],
        "title": "ShotOut"
      },
      "ShotZone": {
        "properties": {
          "zone": {
            "type": "string",
            "title": "Zone"
          },
          "zoneLabel": {
            "type": "string",
            "title": "Zonelabel"
          },
          "shots": {
            "type": "integer",
            "title": "Shots"
          },
          "goals": {
            "type": "integer",
            "title": "Goals"
          },
          "xg": {
            "type": "number",
            "title": "Xg"
          },
          "xgPerShot": {
            "type": "number",
            "title": "Xgpershot"
          }
        },
        "type": "object",
        "required": [
          "zone",
          "zoneLabel",
          "shots",
          "goals",
          "xg",
          "xgPerShot"
        ],
        "title": "ShotZone",
        "description": "Aggregate for one area of the pitch."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}